Marc Behrens message : Dear all,
to be able to trace the changes within the review sheet I would like to propose to save it as Excel 2003 XML format.
It feels like an excel sheet and with 36kilo bytes it is 12 kilo bytes bigger than its excel counterpart but this will make up to when it is rewritten, see example file attached.

Kind regards,

Marc